requirements.txt
langchain
langchain-core
langchain-aws
langgraph
boto3
python-dotenv
streamlit
pillow
supabase
